test: reduce mem even more in captum tests
