+++ .gitignore
@@ -62,6 +62,7 @@ outputs/
 # quanda
 cache/
 /.env
+bin/*
 
 assets/demo/*
 bench_out/*
