+++ tests/explainers/wrappers/test_captum_tracin.py
@@ -1,0 +1,204 @@
+import gc
+
+import pytest
+import torch
+from captum.influence import TracInCP, TracInCPFast
+
+from quanda.explainers.wrappers import (
+    CaptumTracInCP,
+    CaptumTracInCPFast,
+    CaptumTracInCPFastRandProj,
+)
+from quanda.utils.common import get_load_state_dict_func
+
+
+@pytest.mark.explainers
+@pytest.mark.parametrize(
+    "test_id, model, dataset, test_data, test_lables, checkpoints, method_kwargs",
+    [
+        (
+            "mnist",
+            "load_mnist_model",
+            "load_mnist_dataset",
+            "load_mnist_test_samples_1",
+            "load_mnist_test_labels_1",
+            "load_mnist_checkpoints",
+            {
+                "batch_size": 1,
+                "sample_wise_grads_per_batch": False,
+                "loss_fn": torch.nn.CrossEntropyLoss(reduction="none"),
+            },
+        ),
+    ],
+)
+def test_captum_tracincp(
+    test_id,
+    model,
+    dataset,
+    test_data,
+    test_lables,
+    checkpoints,
+    method_kwargs,
+    request,
+):
+    model = request.getfixturevalue(model)
+    dataset = request.getfixturevalue(dataset)
+    test_data = request.getfixturevalue(test_data)
+    test_lables = request.getfixturevalue(test_lables)
+    checkpoints = request.getfixturevalue(checkpoints)[:2]
+
+    explainer_captum = TracInCP(
+        model=model,
+        train_dataset=dataset,
+        checkpoints=checkpoints,
+        checkpoints_load_func=get_load_state_dict_func("cpu"),
+        **method_kwargs,
+    )
+    explanations = explainer_captum.influence(inputs=(test_data, test_lables))
+    del explainer_captum
+    gc.collect()
+
+    explainer = CaptumTracInCP(
+        model=model,
+        train_dataset=dataset,
+        checkpoints=checkpoints,
+        checkpoints_load_func=get_load_state_dict_func("cpu"),
+        device="cpu",
+        **method_kwargs,
+    )
+    explanations_exp = explainer.explain(test_data, test_lables)
+
+    assert torch.allclose(explanations, explanations_exp), (
+        "Training data attributions are not as expected"
+    )
+
+
+@pytest.mark.explainers
+@pytest.mark.parametrize(
+    "test_id, model, checkpoint,dataset, test_data, test_labels, checkpoints, method_kwargs",
+    [
+        (
+            "mnist",
+            "load_mnist_model",
+            "load_mnist_last_checkpoint",
+            "load_mnist_dataset",
+            "load_mnist_test_samples_1",
+            "load_mnist_test_labels_1",
+            "load_mnist_checkpoints",
+            {
+                "batch_size": 1,
+                "loss_fn": torch.nn.CrossEntropyLoss(reduction="sum"),
+                "projection_dim": 5,
+                "seed": 42,
+            },
+        ),
+    ],
+)
+def test_captum_tracincp_fast_rand_proj(
+    test_id,
+    model,
+    checkpoint,
+    dataset,
+    test_data,
+    test_labels,
+    checkpoints,
+    method_kwargs,
+    request,
+    tmp_path,
+    mocker,
+):
+    model = request.getfixturevalue(model)
+    dataset = request.getfixturevalue(dataset)
+    test_data = request.getfixturevalue(test_data)
+    test_labels = request.getfixturevalue(test_labels)
+    checkpoints = request.getfixturevalue(checkpoints)[:2]
+    final_fc_layer = model.fc_3
+
+    n_test = len(test_data)
+    n_train = len(dataset)
+    mock_influence = mocker.patch(
+        "captum.influence.TracInCPFastRandProj.influence",
+        return_value=torch.randn(n_test, n_train),
+    )
+
+    explainer = CaptumTracInCPFastRandProj(
+        model=model,
+        final_fc_layer=final_fc_layer,
+        train_dataset=dataset,
+        checkpoints=checkpoints,
+        checkpoints_load_func=get_load_state_dict_func("cpu"),
+        device="cpu",
+        **method_kwargs,
+    )
+    explanations = explainer.explain(test_data, test_labels)
+
+    mock_influence.assert_called_once()
+    assert explanations.shape == (n_test, n_train)
+
+
+@pytest.mark.explainers
+@pytest.mark.parametrize(
+    "test_id, model, checkpoint,dataset, test_data, test_labels, checkpoints, method_kwargs",
+    [
+        (
+            "mnist",
+            "load_mnist_model",
+            "load_mnist_last_checkpoint",
+            "load_mnist_dataset",
+            "load_mnist_test_samples_1",
+            "load_mnist_test_labels_1",
+            "load_mnist_checkpoints",
+            {
+                "batch_size": 1,
+                "loss_fn": torch.nn.CrossEntropyLoss(reduction="sum"),
+            },
+        ),
+    ],
+)
+def test_captum_tracincp_fast(
+    test_id,
+    model,
+    checkpoint,
+    dataset,
+    test_data,
+    test_labels,
+    checkpoints,
+    method_kwargs,
+    request,
+    tmp_path,
+):
+    model = request.getfixturevalue(model)
+    dataset = request.getfixturevalue(dataset)
+    test_data = request.getfixturevalue(test_data)
+    test_labels = request.getfixturevalue(test_labels)
+    checkpoints = request.getfixturevalue(checkpoints)[:2]
+    final_fc_layer = model.fc_3
+
+    explainer_captum = TracInCPFast(
+        model=model,
+        final_fc_layer=final_fc_layer,
+        train_dataset=dataset,
+        checkpoints=checkpoints,
+        checkpoints_load_func=get_load_state_dict_func("cpu"),
+        **method_kwargs,
+    )
+    explanations = explainer_captum.influence(
+        inputs=(test_data, test_labels), k=None
+    )
+    del explainer_captum
+    gc.collect()
+
+    explainer = CaptumTracInCPFast(
+        model=model,
+        final_fc_layer=final_fc_layer,
+        train_dataset=dataset,
+        checkpoints=checkpoints,
+        checkpoints_load_func=get_load_state_dict_func("cpu"),
+        device="cpu",
+        **method_kwargs,
+    )
+    explanations_exp = explainer.explain(test_data, test_labels)
+
+    assert torch.allclose(explanations, explanations_exp), (
+        "Training data attributions are not as expected"
+    )
